+++ patches/src/main/kotlin/blackroom_vp/SDKTargetVersionPatch.kt
@@ -27,10 +27,23 @@ val mainPatch = resourcePatch(name="SDK 버전 변경",description = "대상 SDK
 	execute()
 	{
 		document("AndroidManifest.xml").use { d ->
-            d.documentElement.setAttribute("platformBuildVersionCode", "30")
             val usesSdk = d.getElementsByTagName("uses-sdk").item(0) as? Element
                 ?: d.createElement("uses-sdk").also { d.documentElement.appendChild(it) }
-            usesSdk.setAttribute("android:targetSdkVersion", "30")
+            usesSdk.setAttribute("android:targetSdkVersion", "28")
+        }
+	}
+}
+
+//Code from https://github.com/rushiranpise/morphe-patches/blob/main/patches/src/main/kotlin/app/template/patches/shared/universal/UniversalPatches.kt#L161
+@Suppress("unused")
+val universalPatch = resourcePatch(name="SDK 버전 변경 (28)",description = "대상 SDK를 강제로 변경하여 Android 12~14 기기에서 설치 가능하도록 조정합니다. (갤23, 폴드/플립5 등)\n64Bit만 지원하는 기기에는 여전히 설치가 어렵습니다 (갤24+, 폴드/플립6+ 등 불가)",default = false)
+{
+	execute()
+	{
+		document("AndroidManifest.xml").use { d ->
+            val usesSdk = d.getElementsByTagName("uses-sdk").item(0) as? Element
+                ?: d.createElement("uses-sdk").also { d.documentElement.appendChild(it) }
+            usesSdk.setAttribute("android:targetSdkVersion", "28")
         }
 	}
 }